requirements.txt
kubernetes==12.0.1
Menu==3.2.2
ruamel-yaml==0.15.46
PyYAML==3.13